refactor(markdown)!: the CST types live in the one public package
The module's CONTRIBUTING.md admits one public package and forbids grouping
types under it by role: a sub-package name joins the compatibility surface, so
moving a type between roles later stops being a refactor and becomes a break.
The CST work added `morphir.langkit.markdown.cst` anyway, holding `MdCstNode`,
`Cst`, `CstParser`, `Lower` and `LinkForm`.

All five move up beside `MdNode`. Nothing public is left behind, so the rest
of the directory goes to `internal` rather than keeping a package alive for it:
`CstFragment`, `Marker`, `InlineSlot`, `CstCollector`, `LinkNote` and
`InlineNotes` are already `private[markdown]`, and AGENTS.md names `internal`
as where machinery belongs. `LinkForm` splits out of `InlineNotes.scala` into
its own file, since the two halves of that file now sit on opposite sides of
the boundary. The directory follows the package rather than diverging from it —
Scala does not require the two to agree, but a `cst/` directory whose files all
declare `package morphir.langkit.markdown` invites the next reader to restore
the package it looks like.

`MD.cst` is untouched. It stays the verb namespace — `parse`, `print`,
`tilingErrors`, `lower` — because it is an object member rather than a package,
so it groups the verbs at a call site without putting `cst` in any type's
fully qualified name. Only the export paths behind it change.

The CONTRIBUTING table now names the CST types on both sides of the boundary,
and says why `MD.cst` is not the thing the rule forbids.

Breaking for anyone who imported `morphir.langkit.markdown.cst`. That package
has never been released.
