+++ morphir/langkit/markdown/CONTRIBUTING.md
@@ -22,8 +22,14 @@ Depth is still available — it just goes downward, not sideways:
 
 | Where | What | Visibility |
 | --- | --- | --- |
-| `morphir.langkit.markdown` | the public surface: `MdNode`, `HeadingLevel`, `FenceInfo`, `Parser`, `ParseError`, `Compiler`, and each writer's entry point | public |
-| `morphir.langkit.markdown.internal` | machinery no caller should name | `private[markdown]` |
+| `morphir.langkit.markdown` | the public surface: `MD`, `dsl`, `MdNode`, `MdCstNode`, `Cst`, `CstParser`, `Lower`, `LinkForm`, `HeadingLevel`, `FenceInfo`, `MdProfile`, `MdStyle`, `MdMeta`, `MdParseError`, `YamlDocText`, `Compiler`, and each writer's entry point | public |
+| `morphir.langkit.markdown.internal` | machinery no caller should name — `Parser`, `MdWriter`, `InlineParser`, `CstFragment`, `InlineNotes` | `private[markdown]` |
+
+The concrete syntax tree is the case worth spelling out, because it is where the rule is easiest to break. `MdCstNode`
+and the three objects over it sit in the root package beside `MdNode`, not in a `cst` package; the fragments and inline
+notes the parser records on the way to building one are `private[markdown]` and sit in `internal`. `MD.cst` is a *verb*
+namespace — an object inside `MD` holding `parse`, `print`, `tilingErrors` and `lower` — which is a member name rather
+than a package name, so it groups the verbs at a call site without putting `cst` in any type's fully qualified name.
 
 **This makes `morphir.langkit.markdown` a split package** — three published artifacts contribute types to it. On a
 classpath that is fine, and it is what makes the single import work. It would not be fine on a Java module path:
