+++ morphir/langkit/markdown/scalatags/test/src/morphir/langkit/markdown/WriterFidelityTests.scala
@@ -2,8 +2,6 @@ package morphir.langkit.markdown
 
 import kyo.*
 import kyo.test.*
-import morphir.langkit.markdown.cst.CstParser
-import morphir.langkit.markdown.cst.Lower
 import morphir.langkit.markdown.dsl.*
 import morphir.langkit.markdown.dsl.given
 import morphir.langkit.markdown.internal.MdWriter
