+++ morphir/langkit/markdown/test/src/morphir/langkit/markdown/FrontmatterTests.scala
@@ -3,9 +3,6 @@ package morphir.langkit.markdown
 import kyo.*
 import kyo.test.*
 import morphir.langkit.core.Span
-import morphir.langkit.markdown.cst.Cst
-import morphir.langkit.markdown.cst.CstParser
-import morphir.langkit.markdown.cst.MdCstNode
 import morphir.langkit.markdown.internal.Parser
 
 /**
