+++ morphir/langkit/markdown/test/src/morphir/langkit/markdown/MdWriterTests.scala
@@ -2,7 +2,6 @@ package morphir.langkit.markdown
 
 import kyo.*
 import kyo.test.*
-import morphir.langkit.markdown.cst.Cst
 import morphir.langkit.markdown.dsl.*
 import morphir.langkit.markdown.dsl.given
 import morphir.langkit.markdown.internal.{MdWriter, Parser}
