+++ morphir/langkit/markdown/trees/test/src/morphir/langkit/markdown/trees/MarkdownTreesTests.scala
@@ -4,7 +4,7 @@ import kyo.*
 import kyo.test.*
 import morphir.langkit.markdown.{MdNode, MdProfile}
 import morphir.langkit.markdown.internal.Parser
-import morphir.langkit.markdown.cst.{Cst, MdCstNode, CstParser}
+import morphir.langkit.markdown.{Cst, MdCstNode, CstParser}
 import morphir.langkit.trees.{NodeTypeName, QueryableTree}
 import morphir.langkit.trees.unist.UnistProjection
